Update: move Empty to component
